Added `IconAwareTrait`
