--- test/functional/AbstractBaseMenuItemTest.php
@@ -1,60 +1,0 @@
-<?php
-
-namespace RebelCode\WordPress\Admin\Menu\FuncTest;
-
-use Xpmock\TestCase;
-use RebelCode\WordPress\Admin\Menu\AbstractBaseMenuItem;
-
-/**
- * Tests {@see RebelCode\WordPress\Admin\Menu\AbstractBaseMenuItem}.
- *
- * @since [*next-version*]
- */
-class AbstractBaseMenuItemTest extends TestCase
-{
-    /**
-     * The classname of the test subject.
-     *
-     * @since [*next-version*]
-     */
-    const TEST_SUBJECT_CLASSNAME = 'RebelCode\\WordPress\\Admin\\Menu\\AbstractBaseMenuItem';
-
-    /**
-     * Creates a new instance of the test subject.
-     *
-     * @since [*next-version*]
-     */
-    public function createInstance()
-    {
-        $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
-            ->_onSelected()
-            ->new();
-
-        return $mock;
-    }
-
-    /**
-     * Tests whether a valid instance of the test subject can be created.
-     *
-     * @since [*next-version*]
-     */
-    public function testCanBeCreated()
-    {
-        $subject = $this->createInstance();
-
-        $this->assertInstanceOf(
-            static::TEST_SUBJECT_CLASSNAME, $subject,
-            'Subject is not a valid instance'
-        );
-
-        $this->assertInstanceOf(
-            'RebelCode\\WordPress\\Admin\\Menu\\MenuItemInterface', $subject,
-            'Subject does not implement expected interface.'
-        );
-
-        $this->assertInstanceOf(
-            'RebelCode\\WordPress\\Admin\\Menu\\MenuElementInterface', $subject,
-            'Subject does not implement expected interface.'
-        );
-    }
-}
